requirements.txt
# ── Entorno ───────────────────────────────────────────────────────────────────
python-dotenv==1.0.1

# ── Base de datos ─────────────────────────────────────────────────────────────
sqlalchemy==2.0.36
psycopg2-binary==2.9.10

# ── ETL y manipulacion de datos ───────────────────────────────────────────────
pandas==2.2.3
numpy==1.26.4
requests==2.32.3
APScheduler==3.10.4

# ── Imputacion ────────────────────────────────────────────────────────────────
lightgbm==4.5.0

# ── Modelos estadisticos ──────────────────────────────────────────────────────
statsmodels==0.14.4

# ── Modelos de machine learning ───────────────────────────────────────────────
xgboost==2.1.3
scikit-learn==1.5.2

# ── Deep learning ─────────────────────────────────────────────────────────────
tensorflow==2.17.0

# ── Dashboard ─────────────────────────────────────────────────────────────────
streamlit==1.40.2
plotly==5.24.1

# ── Variables exogenas externas ───────────────────────────────────────────────
holidays==0.59
fredapi==0.5.1

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.3.4
pytest-cov==6.0.0
